requirements.txt
microsoft-graph-core
azure-ai-textanalytics
openai